Changed Vector2 back to floats; Renderable now keeps an internal position and casts back to int just before rendering

+++ Source/Vector2.h
@@ -8,7 +8,7 @@ public:
     {
     }
 
-    Vector2(unsigned int _x, unsigned int _y)
+    Vector2(float _x, float _y)
         : x(_x)
         , y(_y)
     {
@@ -24,6 +24,6 @@ public:
         return false;
     }
 
-    unsigned int x;
-    unsigned int y;
+    float x;
+    float y;
 };
